requirements.txt
ipdb
ipython
pandas
python-dotenv
python-decouple

